Hack together complete module restructure into several submodules.
