+++ crates/bevy_input/src/abstract_device/abstract_input_device.rs
@@ -1,0 +1,9 @@
+use crate::core::{Axis, Input};
+use bevy_app::{EventReader, Events};
+use bevy_ecs::{Local, Res, ResMut};
+use bevy_utils::HashMap;
+
+/// Input device container
+#[derive(Debug, Copy, Clone, PartialEq, Eq, Hash)]
+#[cfg_attr(feature = "serialize", derive(serde::Serialize, serde::Deserialize))]
+pub struct AbstractInputDevice(pub usize);
