+++ crates/bevy_input/src/abstract_device/abstract_input_device_plugin.rs
@@ -1,0 +1,1 @@
+
