+++ crates/bevy_input/src/abstract_device/mod.rs
@@ -1,0 +1,10 @@
+pub mod abstract_input_device;
+pub mod abstract_input_device_plugin;
+
+// export public exports module
+pub use prelude::*;
+
+/// Public exports module
+pub mod prelude {
+    pub use super::abstract_input_device::AbstractInputDevice;
+}
