+++ crates/bevy_input/src/touch/mod.rs
@@ -3,6 +3,9 @@ use bevy_ecs::{Local, Res, ResMut};
 use bevy_math::Vec2;
 use bevy_utils::{HashMap, HashSet};
 
+pub mod touchplugin;
+pub use touchplugin::TouchInputPlugin;
+
 /// A touch input event
 #[derive(Debug, Clone)]
 pub struct TouchInput {
